[#29] refactor: nav 영역
why: 미세먼지 정보 창과 예보창이 spa방식으로 리팩토링을 진행했기에,
nav영역도 마찬가지 방식으로 교체했다.
