+++ client-web/src/nav/index.js
@@ -1,0 +1,9 @@
+import { displayTab, forecastTab } from "./view";
+import css from "./style/nav.css";
+
+const mount = target => {
+  target.appendChild(displayTab());
+  target.appendChild(forecastTab());
+};
+
+mount(document.querySelector(".dust-nav"));
