+++ client-web/src/nav/style/nav.css
@@ -1,0 +1,40 @@
+.dust-nav {
+  display: flex;
+}
+
+.dust-tab {
+  border: 1px solid #000;
+  height: 5vh;
+  width: 50vw;
+  font-size: 1.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.selected {
+  background-color: #37efe3;
+  color: #eaf7f5;
+}
+
+#app {
+  position: relative;
+  height: 90vh;
+}
+
+#dust-display {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  z-index: 1;
+}
+
+#dust-forecast {
+  visibility: hidden;
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100vw;
+  z-index: 0;
+}
