--- client-web/style/display.css
@@ -1,50 +1,0 @@
-body {
-  box-sizing: border-box;
-}
-
-.dust-status-display {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  height: 40vh;
-  background: linear-gradient(
-    180deg,
-    rgba(61, 172, 247, 1),
-    rgba(61, 172, 247, 0)
-  );
-}
-
-.header {
-  font-size: 2rem;
-  text-align: center;
-}
-
-.dust-status-face {
-  display: flex;
-  justify-content: center;
-}
-
-.dust-status-text {
-  font-size: 1.5rem;
-  font-weight: 900;
-  text-align: center;
-}
-
-.dust-status-today {
-  display: flex;
-  width: 100vw;
-  justify-content: space-around;
-  font-size: 1.2rem;
-  font-weight: 400;
-}
-
-.dust-observation-station {
-  text-align: center;
-  font-size: 0.7rem;
-  font-weight: 400;
-}
-
-.dust-chart {
-  height: 50vh;
-  overflow: scroll;
-}
