--- client-web/style/swipe.css
@@ -1,23 +1,0 @@
-#app {
-  position: relative;
-  max-height: 100vh;
-}
-
-#dust-display {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 1;
-}
-
-#dust-forecast {
-  visibility: hidden;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  z-index: 0;
-}
